We now have some of the functionality demonstrated to render great sex option screens in MPLAT. Needs some love.
Also improved auto-configuration of the NPP plugin installer.
Increase max texture size in MPLAT to 40964096

Moved demo optiions to great.sex.test-data.cpp
Corrected case of Yes.tiff in the game options bool thing
et al
