+++ source/rococo/include/rococo.gr.client.h
@@ -55,8 +55,6 @@ namespace Rococo
 			// Triggers WM_PAINT as soon as the message queue is free
 			virtual void QueuePaint() = 0;
 
-			virtual void ShowError(Vec2i start, Vec2i end, cstr nameRef, cstr sourceBuffer, cstr message) = 0;
-
 			virtual [[nodiscard]] IGRSystem& GRSystem() = 0;
 
 			// Tries to load a frame specified by the great sex sexml File into the parent widget.
