+++ source/rococo/rococo.gui.retained/rococo.gr.game-options.bool.cpp
@@ -47,8 +47,8 @@ namespace GRANON
 			button->Widget().Panel().SetExpandToParentHorizontally();
 			button->Widget().Panel().SetExpandToParentVertically();
 			button->MakeToggleButton();
-			button->SetPressedImagePath("!textures/toolbars/3rd-party/www.aha-soft.com/yes.tiff");
-			button->SetRaisedImagePath("!textures/toolbars/3rd-party/www.aha-soft.com/no.tiff");
+			button->SetPressedImagePath("!textures/toolbars/3rd-party/www.aha-soft.com/Yes.tiff");
+			button->SetRaisedImagePath("!textures/toolbars/3rd-party/www.aha-soft.com/No.tiff");
 			button->Panel().Remove(EGRPanelFlags::AcceptsFocus);
 			
 			MakeTransparent(button->Widget().Panel(), EGRSchemeColourSurface::BUTTON);
