+++ source/rococo/rococo.mhost/mhost.app.cpp
@@ -29,6 +29,7 @@ using namespace Rococo::Graphics::Textures;
 using namespace Rococo::Script;
 using namespace MHost::OS;
 using namespace Rococo::Sex;
+using namespace Rococo::GreatSex;
 
 struct AppArgs
 {
@@ -205,7 +206,7 @@ namespace MHost
 		public Strings::IStringPopulator,
 		public Rococo::MPEditor::IMPEditorEventHandler,
 		IO::IShaderMonitorEventHook,
-		public GreatSex::IGreatSexResourceLoader
+		public IGreatSexResourceLoader
 	{
 		Platform& platform;
 		IDirectAppControl& control;
@@ -223,7 +224,6 @@ namespace MHost
 		int32 guiToggleKey = 0;
 
 		AutoFree<IPackageSupervisor> packageMHost;
-		AutoFree<GreatSex::IGreatSexGeneratorSupervisor> greatSex;
 
 		// IO::IShaderMonitorEventHook
 		void OnLog(IO::IShaderMonitor& monitor, IO::EShaderLogPriority priority, cstr file, cstr message) override
@@ -351,7 +351,6 @@ namespace MHost
 		App(Platform& _platform, IDirectAppControl& _control, const AppArgs& args) :
 			platform(_platform), control(_control), sceneManager(_platform)
 		{
-			greatSex = GreatSex::CreateGreatSexGenerator(Rococo::Memory::CheckedAllocator(), *this);
 			busyPanel = platform.graphics.gui.BindPanelToScript("!scripts/panel.opening.sxy", nullptr, Rococo::NoImplicitIncludes());
 
 			platform.plumbing.publisher.Subscribe(this, Rococo::Events::evBusy);
@@ -386,6 +385,47 @@ namespace MHost
 		void Populate(cstr) override
 		{
 			platform.os.ios.EnumerateModifiedFiles(*this);
+		}
+
+		void LoadFrame(cstr sexmlPath, Gui::IGRWidgetMainFrame& frame, IEventCallback<IGreatSexGenerator>& onConstruct)
+		{
+			struct ErrorHandler : IEventCallback<LoadFrameException>
+			{
+				void OnEvent(LoadFrameException& ex) override
+				{
+					Windows::ShowErrorBox(p->os.mainWindow, ex.errorCode, ex.message, ex.filename);
+				}
+
+				Platform* p = nullptr;
+			} onError;
+
+			onError.p = &platform;
+
+			GreatSex::LoadFrame(platform.os.installation, Memory::CheckedAllocator(), sexmlPath, frame, onConstruct, onError);
+		}
+
+		void SetGreatSexOverlay(const fstring& pingPath) override
+		{
+			SetEditorVisibility(true);
+			auto* frame = platform.graphics.GR.Root().GR().FindFrame(ID_EDITOR_FRAME);
+			if (frame)
+			{
+				frame->MenuBar().Panel().Parent()->SetCollapsed(true);
+
+				auto& scheme = frame->Widget().Panel().Root().Scheme();
+				Gui::SetSchemeColours_ThemeGrey(scheme);
+				Gui::SetPropertyEditorColours_PastelScheme(frame->Panel());
+
+				struct : IEventCallback<GreatSex::IGreatSexGenerator>
+				{
+					void OnEvent(GreatSex::IGreatSexGenerator& generator) override
+					{
+						GreatSex::AddTestOptions(generator);
+					}
+				} onConstruct;
+
+				LoadFrame(pingPath, *frame, onConstruct);
+			}
 		}
 
 		std::vector<HString> messages;
@@ -603,7 +643,21 @@ namespace MHost
 						AppendToMenuRecursive(*frame, idProfileMenu, subsystem, id);
 					}
 				);
+
+				LoadTestPage(*frame);
 			}
+		}
+
+		void LoadTestPage(Gui::IGRWidgetMainFrame& frame)
+		{
+			struct : IEventCallback<IGreatSexGenerator>
+			{
+				void OnEvent(IGreatSexGenerator& generator)
+				{
+					GreatSex::AddTestOptions(generator);
+				}
+			} onConstruct;
+			LoadFrame("!tests/greatsex.test.sexml", frame, onConstruct);
 		}
 
 		void ClearMenus() override
@@ -988,6 +1042,7 @@ namespace MHost
 
 			platform.graphics.GR.GarbageCollect();
 
+			AutoFree<IGreatSexGeneratorSupervisor> greatSex = CreateGreatSexGenerator(Memory::CheckedAllocator(), *this);
 			greatSex->AppendWidgetTreeFromSexML(s, frame->ClientArea().Widget());
 		}
 	};
